+++ src/app/commands/track.go
@@ -4,7 +4,7 @@ import (
 	. "github.com/jotaen/klog/src"
 	"github.com/jotaen/klog/src/app"
 	"github.com/jotaen/klog/src/app/lib"
-	"github.com/jotaen/klog/src/parser/reconciler"
+	"github.com/jotaen/klog/src/parser/reconciling"
 	"strings"
 )
 
@@ -28,24 +28,20 @@ func (opt *Track) Run(ctx app.Context) error {
 	opt.NoStyleArgs.Apply(&ctx)
 	date := opt.AtDate(ctx.Now())
 	value := sanitiseQuotedLeadingDash(opt.Entry)
-	return ctx.ReconcileFile(opt.OutputFileArgs.File,
-		func(base reconciler.Reconciler) (*reconciler.ReconcileResult, error) {
-			entryReconciler := reconciler.NewEntryReconciler(base, func(r Record) bool {
-				return r.Date().IsEqualTo(date)
-			})
-			if entryReconciler == nil {
-				return nil, reconciler.NotEligibleError{}
-			}
-			return entryReconciler.AppendEntry(func(r Record) string { return value })
+	return ctx.ReconcileFile(
+		opt.OutputFileArgs.File,
+		func(reconciler reconciling.Reconciler) (*reconciling.Result, error) {
+			return reconciler.AppendEntry(
+				func(r Record) bool { return r.Date().IsEqualTo(date) },
+				func(r Record) string { return value })
 		},
-		func(base reconciler.Reconciler) (*reconciler.ReconcileResult, error) {
-			recordReconciler := reconciler.NewRecordReconciler(base, date)
+		func(reconciler reconciling.Reconciler) (*reconciling.Result, error) {
 			headline := opt.AtDate(ctx.Now()).ToString()
-			lines := []reconciler.InsertableText{
+			lines := []reconciling.InsertableText{
 				{headline, 0},
 				{value, 1},
 			}
-			return recordReconciler.InsertBlock(lines)
+			return reconciler.InsertBlock(date, lines)
 		},
 	)
 }
