+++ src/parser/reconciling/template.go
@@ -1,4 +1,4 @@
-package reconciler
+package reconciling
 
 import (
 	"errors"
