+++ src/parser/reconciling/template_test.go
@@ -1,4 +1,4 @@
-package reconciler
+package reconciling
 
 import (
 	"github.com/stretchr/testify/assert"
